Create BlackjackPlayer
This is a Blackjack player that I made to play against my professor's Blackjack dealer. There are small things such as split not being a part of my code yet that I soon plan to adjust on here as time goes on and school becomes less. The code isn't perfect, but as a student learning Java this is my attempt.
Code language: Java